+++ tests/testit/test-other.R
@@ -39,7 +39,7 @@ set(pkgDT, NULL, "versionSpec", NA)
 rmDuplicatePkgs(pkgDT)
 
 out <- detachAll("data.table")
-testit::assert(isTRUE(out == 1))
+testit::assert(isTRUE(out['data.table'] == 1))
 
 warn <- tryCatch(warningCantInstall("devtolls"), warning = function(w) w$message)
 testit::assert(grepl("you will likely", warn))